# Conflicts:
#	SPD-classes/src/main/java/com/watabou/noosa/Camera.java
#	build.gradle
#	core/src/main/assets/consumable_icons.png
#	core/src/main/assets/items.png
#	core/src/main/assets/terrain_features.png
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/Preferences.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/ShatteredPixelDungeon.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/Char.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/Generator.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/Heap.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/armor/Armor.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/armor/ClassArmor.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/armor/glyphs/Entanglement.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/bombs/Bomb.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/food/ChargrilledMeat.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/food/FrozenCarpaccio.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/food/MysteryMeat.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/food/OverpricedRation.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/wands/WandOfFireblast.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/wands/WandOfFrost.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/wands/WandOfPrismaticLight.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/enchantments/Lucky.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/enchantments/Vampiric.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/RegularLevel.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/Room.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/features/Chasm.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/painters/StandardPainter.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/scenes/CellSelector.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/scenes/WelcomeScene.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/sprites/ItemSpriteSheet.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/sprites/WarlockSprite.java
#	core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/windows/WndSettings.java
#	core/src/main/resources/com/shatteredpixel/shatteredpixeldungeon/messages/items/items.properties
#	core/src/main/resources/com/shatteredpixel/shatteredpixeldungeon/messages/windows/windows.properties

+++ SPD-classes/src/main/java/com/watabou/gltextures/SmartTexture.java
@@ -3,7 +3,7 @@
  * Copyright (C) 2012-2015  Oleg Dolya
  *
  * Shattered Pixel Dungeon
- * Copyright (C) 2014-2016 Evan Debenham
+ * Copyright (C) 2014-2017 Evan Debenham
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -119,11 +119,11 @@ public class SmartTexture extends Texture {
 		bitmap = null;
 	}
 	
-	public RectF uvRect( int left, int top, int right, int bottom ) {
+	public RectF uvRect( float left, float top, float right, float bottom ) {
 		return new RectF(
-			(float)left		/ width,
-			(float)top		/ height,
-			(float)right	/ width,
-			(float)bottom	/ height );
+			left		/ width,
+			top		/ height,
+			right	/ width,
+			bottom	/ height );
 	}
 }
